+++ src/components/Home.js
@@ -1,12 +1,11 @@
 import React, {Component} from 'react';
-import ItemsList from '../ItemsList.js';
-import Filtros from '../Filtros.js';
+import ItemsList from './ItemsList.js';
 import {NavLink} from 'react-router-dom';
-import '../../App.css';
-import MarvApi from '../services/MarvApi.js';
+import Comics from './Comics.js';
+import MarvApi from './services/MarvApi.js';
 
 
-class Comics extends Component {
+class Home extends Component {
 	constructor(props) {
 		super(props);
 		this.state = {
@@ -36,33 +35,15 @@ class Comics extends Component {
 			})
 		})
 	}
-	componentWillUnmount(){
 
-	}
 
 	render(){
-		const {error,loading,comics,characters} = this.state
-		{console.log('new', this.state.comics)}
-		return(
-				<div>
+		const {comics,characters} = this.state
+		return (  
+			<div>
 				<main role="main">
 				    <div className="py-5 bg-light">
 				        <div className="container">
-				        <h2>Comics </h2>
-				        {/*Botones filtro*/}
-				        <form className="form-inline my-2 my-md-0" action="">
-				        	<select className="form-control"> 
-				        		<option>Año</option>
-				        	</select>
-				        	<select className="form-control">
-				        		<option>Ordenado por</option>
-				        	</select>
-				        	<select className="form-control">
-				        		<option>Generos</option>
-				        	</select>
-				        <Filtros />
-                		</form>
-
 				            <section className="items-section">
 				                <div className="items-section-body">
 				                    <div className="row">
@@ -70,12 +51,19 @@ class Comics extends Component {
 				                    </div>
 				                </div>
 				            </section>
-			            
+				            <section className="items-section">
+				                <h5 className="items-section-title">Personajes más Populares <NavLink to="/characters">Ver todas</NavLink></h5>
+				                <div className="items-section-body">
+				                    <div className="row">
+				                        <ItemsList listadoResultados={characters} />
+				                    </div>
+				                </div>
+				            </section>
 				        </div>
 				    </div>
 				</main>
 			</div>
-			)
-	}
+)
 }
-export default Comics;
+}
+export default Home;
